flatten modules
