--- functions-nest/src/resources/attachment/controllers/index.ts
@@ -1,1 +1,0 @@
-export * from './attachment.controller'
