--- functions-nest/src/resources/attachment/services/index.ts
@@ -1,1 +1,0 @@
-export * from './attachment.service'
