+++ functions-nest/src/resources/clothing/clothing.module.ts
@@ -2,8 +2,8 @@ import { MikroOrmModule } from '@mikro-orm/nestjs'
 import { Logger, Module } from '@nestjs/common'
 
 import { Clothing } from './models'
-import { ClothingService } from './services'
-import { ClothingController } from './controllers'
+import { ClothingService } from './clothing.service'
+import { ClothingController } from './clothing.controller'
 
 @Module({
 	imports: [MikroOrmModule.forFeature([Clothing])],
