--- functions-nest/src/resources/clothing/controllers/index.ts
@@ -1,1 +1,0 @@
-export * from './clothing.controller'
