+++ functions-nest/src/resources/clothing/models.ts
