--- functions-nest/src/resources/clothing/services/index.ts
@@ -1,1 +1,0 @@
-export * from './clothing.service'
