--- functions-nest/src/resources/gear/controllers/index.ts
@@ -1,1 +1,0 @@
-export * from './gear.controller'
