+++ functions-nest/src/resources/gear/gear.module.ts
@@ -2,8 +2,8 @@ import { MikroOrmModule } from '@mikro-orm/nestjs'
 import { Logger, Module } from '@nestjs/common'
 
 import { Gear } from './models'
-import { GearService } from './services'
-import { GearController } from './controllers'
+import { GearService } from './gear.service'
+import { GearController } from './gear.controller'
 
 @Module({
 	imports: [MikroOrmModule.forFeature([Gear])],
