--- functions-nest/src/resources/gear/services/index.ts
@@ -1,1 +1,0 @@
-export * from './gear.service'
