--- functions-nest/src/resources/loadout/controllers/index.ts
@@ -1,1 +1,0 @@
-export * from './loadout.controller'
