+++ functions-nest/src/resources/loadout/loadout.module.ts
@@ -2,8 +2,8 @@ import { MikroOrmModule } from '@mikro-orm/nestjs'
 import { Logger, Module } from '@nestjs/common'
 
 import { Loadout } from './models'
-import { LoadoutService } from './services'
-import { LoadoutController } from './controllers'
+import { LoadoutService } from './loadout.service'
+import { LoadoutController } from './loadout.controller'
 
 @Module({
 	imports: [MikroOrmModule.forFeature([Loadout])],
