+++ functions-nest/src/resources/loadout/loadout.service.ts
@@ -1,19 +1,19 @@
 import { Injectable } from '@nestjs/common'
+import { InjectRepository } from '@mikro-orm/nestjs'
 import { EntityRepository, QueryOrder } from '@mikro-orm/core'
 
 import { FirebaseUserService } from 'src/firebase/services/firebase-user.service'
-import { InjectRepository } from '@mikro-orm/nestjs'
-import { Weapon } from '../models'
+import { Loadout } from './models'
 
 @Injectable()
-export class WeaponService {
+export class LoadoutService {
 	constructor(
-		@InjectRepository(Weapon)
-		private repo: EntityRepository<Weapon>,
+		@InjectRepository(Loadout)
+		private repo: EntityRepository<Loadout>,
 		private user: FirebaseUserService,
 	) {}
 
-	async getAll(): Promise<Weapon[]> {
+	async getAll(): Promise<Loadout[]> {
 		// TODO: Remove uid from result
 		const result = await this.repo.find(
 			{
@@ -21,13 +21,33 @@ export class WeaponService {
 			},
 			{
 				orderBy: { createdAt: QueryOrder.ASC },
+				populate: { weapons: true },
 			},
 		)
 
 		return result
 	}
 
-	async add(dto: any): Promise<Weapon> {
+	async getById(id: string): Promise<Loadout> {
+		// TODO: Remove uid from result
+		const result = await this.repo.findOne(
+			{
+				id,
+				$or: [{ uid: this.user.uid }, { shared: true }],
+			},
+			{
+				weapons: {
+					attachments: true,
+				},
+				clothing: true,
+				gear: true,
+			},
+		)
+
+		return result
+	}
+
+	async add(dto: any): Promise<Loadout> {
 		// TODO: Validate no id being sent in at controller
 		const newEntity = this.repo.create({ uid: this.user.uid, ...dto })
 		this.repo.persistAndFlush(newEntity)
