+++ functions-nest/src/resources/loadout/models.ts
