--- functions-nest/src/resources/loadout/services/index.ts
@@ -1,1 +1,0 @@
-export * from './loadout.service'
