--- functions-nest/src/resources/weapon/controllers/index.ts
@@ -1,1 +1,0 @@
-export * from './weapon.controller'
