+++ functions-nest/src/resources/weapon/models.ts
