--- functions-nest/src/resources/weapon/services/index.ts
@@ -1,1 +1,0 @@
-export * from './weapon.service'
