+++ functions-nest/src/resources/weapon/weapon.controller.spec.ts
