+++ functions-nest/src/resources/weapon/weapon.module.ts
@@ -1,9 +1,9 @@
-import { MikroOrmModule } from '@mikro-orm/nestjs'
 import { Logger, Module } from '@nestjs/common'
+import { MikroOrmModule } from '@mikro-orm/nestjs'
 
 import { Weapon } from './models'
-import { WeaponService } from './services'
-import { WeaponController } from './controllers/'
+import { WeaponService } from './weapon.service'
+import { WeaponController } from './weapon.controller'
 
 @Module({
 	imports: [MikroOrmModule.forFeature([Weapon])],
